Panel for going to a specific point added. Loading last used anchors feature added

+++ main.py
@@ -1,11 +1,12 @@
 """The main file that creates an application to use the microscope"""
 
-from tkinter import Tk, Frame, mainloop, messagebox
+from tkinter import Tk, Frame, mainloop, messagebox, Menu, Toplevel
 import logging
 
 from modules.classes.scene_switcher import SceneSwitcher
 from modules.classes.logger import Logger
-from modules.guis import session_loader, connection, anchors
+from modules.classes.session_data import data_struct, load_last_anchors
+from modules.guis import session_loader, connection, anchors, go_to_nth
 from modules.guis.PointRecording import point_record
 from modules.guis.mainMenu import main_scene
 from modules.helpers.configuration import (
@@ -28,16 +29,40 @@ if not LOOPBACK_EXISTS:
         "If you wish to connect to SOLIS, fix your connection and restart the program.")
 
 #set up the config file for SOLIS
-with open("./SOLIS.cfg","w",encoding='utf-8') as f:
-    f.write(f"STAGE_PORT\n{STAGE_PORT[3:]}\nAPP_PORT\n{LOOPBACK_B[3:]}\nEND")
-log.info("Ports set. Loopback ports: %s and %s, Stage port: %s", LOOPBACK_A, LOOPBACK_B, STAGE_PORT)
+if STAGE_EXISTS and LOOPBACK_EXISTS:
+    with open("./SOLIS.cfg","w",encoding='utf-8') as f:
+        f.write(f"STAGE_PORT\n{STAGE_PORT[3:]}\nAPP_PORT\n{LOOPBACK_B[3:]}\nEND")
+    log.info("Ports set. Loopback ports: %s and %s, Stage port: %s", LOOPBACK_A, LOOPBACK_B, STAGE_PORT)
+else:
+    log.warning("Ports not available.")
 
+def open_move_to_nth_point() -> None:
+    """Opens a toplevel that enables moving to a specific point"""
+    if (data_struct.anchors_set and
+        data_struct.local_anchors_set and
+        data_struct.points_set):
+
+        log.info("Opening Nth point window")
+        top: Toplevel=Toplevel()
+        go_to_nth.GUI(top).pack()
+
+    else:
+        messagebox.showwarning("Unable to open window.",#type: ignore
+            "Points and/or anchors have not been set.")
 
 # creating main tkinter window/toplevel
 master: Tk = Tk()
 master.title("SolisXY GUI")
 master.grid_rowconfigure(1, weight=1)
 master.grid_columnconfigure(1, weight=1)
+
+menubar: Menu = Menu(master)
+filemenu: Menu = Menu(menubar, tearoff=0)
+filemenu.add_command(label="Load last anchors", command=load_last_anchors)
+filemenu.add_command(label="Move to N-th point", command=open_move_to_nth_point)
+filemenu.add_separator()
+filemenu.add_command(label="Exit", command=master.quit)
+menubar.add_cascade(label="Options", menu=filemenu)
 
 connection_frame: Frame=connection.GUI(master)
 connection_frame.grid(row=0,column=0)
@@ -95,4 +120,5 @@ master.update()
 
 # infinite loop which can be terminated by keyboard
 # or mouse interrupt
+master.config(menu=menubar)
 mainloop()
